requirements.txt
# Core dependencies
psycopg2-binary>=2.9.9
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.3

# Development
pytest>=7.4.3
